+++ 2_module_parser/env_handler1.c
@@ -1,7 +1,16 @@
-
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_handler1.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vsimeono <vsimeono@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/07/19 09:29:47 by vsimeono          #+#    #+#             */
+/*   Updated: 2022/07/26 12:29:56 by vsimeono         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "parser.h"
-
 /* Find PWD from the ENV, if not there it Get's it from Memory, Return NULL if Error */
 static char	*get_pwd(t_envp_data envp_data)
 {
@@ -107,27 +116,6 @@ int	env_resolver(t_data *data, char **input)
 	return (1);
 }
 
-/* Return New Allocated String, Old Substring Length with new Substring at Index */
-char	*str_replace_str_at(char *str, int idx, int length, char *replacement)
-{
-	char	*new_str;
-	char	*tmp;
 
-	if (!str || idx < 0 || length < 0 || !replacement)
-		return (NULL);
-	if (idx >= (int)ft_strlen(str))
-		return (ft_strjoin(str, replacement));
-	if (length > (int)ft_strlen(&str[idx]))
-		length = (int)ft_strlen(&str[idx]);
-	str[idx] = '\0';
-	tmp = ft_strjoin(str, replacement);
-	if (!tmp)
-		return (NULL);
-	new_str = ft_strjoin(tmp, &str[idx + length]);
-	free(tmp);
-	if (!new_str)
-		return (NULL);
-	return (new_str);
-}
