+++ 2_module_parser/env_handler2.c
@@ -1,0 +1,36 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_handler2.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vsimeono <vsimeono@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/07/20 14:20:40 by vsimeono          #+#    #+#             */
+/*   Updated: 2022/07/26 12:30:02 by vsimeono         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parser.h"
+
+/* Return New Allocated String, Old Substring Length with new Substring at Index */
+char	*str_replace_str_at(char *str, int idx, int length, char *replacement)
+{
+	char	*new_str;
+	char	*tmp;
+
+	if (!str || idx < 0 || length < 0 || !replacement)
+		return (NULL);
+	if (idx >= (int)ft_strlen(str))
+		return (ft_strjoin(str, replacement));
+	if (length > (int)ft_strlen(&str[idx]))
+		length = (int)ft_strlen(&str[idx]);
+	str[idx] = '\0';
+	tmp = ft_strjoin(str, replacement);
+	if (!tmp)
+		return (NULL);
+	new_str = ft_strjoin(tmp, &str[idx + length]);
+	free(tmp);
+	if (!new_str)
+		return (NULL);
+	return (new_str);
+}
