+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: vsimeono <vsimeono@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/26 19:47:23 by vsimeono          #+#    #+#             */
-/*   Updated: 2022/07/23 00:53:00 by vsimeono         ###   ########.fr       */
+/*   Updated: 2022/07/26 12:58:48 by vsimeono         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
